Average Perceptron POS Tagger (Issue #2)
I added two files "averagePerceptron.jl" and "tagger.jl" to the "src" folder in "TextAnalysis" , former file contains the Average Perceptron model and latter file is used to train a model, predict tags or reuse a model using model in former file. I have not made any changes to any of the original files so currently the model can't be used through "TextAnalysis.jl", because I in tagger I have to include an optional argument "load" which specifies whether user want to load pretrain weights or not. If the user want to load pretrained weights then the input should "true" (Boolean), which I don't know how should I take (Since I am little new to julia).
I have commented the code for "loadModel" function in "tagger.jl" which should run if user gives that argument as true.
But for now we can train and test the model.
I will give the pretrained weights shortly.
Please suggest me changes to make the better and a way to take that argument("load") from the user
If I am able to take that argument the model will be completed.