Resolving issue not retrieving set for loading pretained model
The problem that we could not load `classes` from the file which contains pretrained weights, tagdict and classes , since `classes` was saved as `Set`. Because there is some issue in BSON , so that it can't load `Set` variables properly. Now the classes are first converted to `Array` and then saved. And while load it is loaded as `Array` and then converted to `Set`.